+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-08-28
+
 ### Added
 
 - Add renovate configuration
@@ -73,7 +75,8 @@ and this project adheres to
 - add django_peertube_runner_connector app
 - add basic quality tools + testing suite 
 
-[unreleased]: https://github.com/openfun/django-peertube-runner-connector/compare/v0.6.0...main
+[unreleased]: https://github.com/openfun/django-peertube-runner-connector/compare/v0.7.0...main
+[0.7.0]: https://github.com/openfun/django-peertube-runner-connector/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/openfun/django-peertube-runner-connector/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/openfun/django-peertube-runner-connector/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/openfun/django-peertube-runner-connector/compare/v0.3.1...v0.4.0